Move styled components from App into their own folders in Elements
