+++ src/Elements/NavLinks/index.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const NavLinks = styled.div`
+  padding-bottom: 25px;
+
+  a {
+    margin-right: 10px;
+    color: white;
+    text-decoration: none;
+  }
+`;
